changed long long to int

+++ utils.cpp
@@ -100,16 +100,16 @@ void flow(int n , int m , int k , vector<vector<int>>& clauses, set<pair<int,int
     }
 }
 //U - 0, R - 1, D- 2, L - 3
-void limit_on_turns(long long n , long long m , long long k , long long J , vector<vector<long long>>& fin_clauses, int seed) {
+void limit_on_turns(int n , int m , int k , int J , vector<vector<int>>& fin_clauses, int seed) {
     
     for(int metro= 0; metro<k ; metro++){
-        vector<long long> variables;
+        vector<int> variables;
         for(int i= 0; i<n; i++){
             for(int j =0; j<m; j++){
                 // (x-1,y, D)^(x,y,R)=>T(metro,x,y) , (x-1,y,D)^(x,y,L)=>T(metro,x,y)
                 variables.push_back(hash_t(metro, i, j, n, m, k));
                 if(i-1>=0){
-                    vector<long long> res;
+                    vector<int> res;
                     res = {-hash_d(metro, i-1, j, 2, n, m, k), -hash_d(metro, i, j, 1, n, m, k), hash_t(metro, i, j, n, m, k)};
                     fin_clauses.push_back(res);
                     res = {-hash_d(metro, i-1, j, 2, n, m, k), -hash_d(metro, i, j, 3, n, m, k), hash_t(metro, i, j, n, m, k)};
@@ -117,7 +117,7 @@ void limit_on_turns(long long n , long long m , long long k , long long J , vect
                 }
                 if(i+1<n){
                     // (x+1,y, U)^(x,y,R)=>T(metro,x,y) , (x+1,y,U)^(x,y,L)=>T(metro,x,y)
-                    vector<long long> res;
+                    vector<int> res;
                     res = {-hash_d(metro, i+1, j, 0, n, m, k), -hash_d(metro, i, j, 1, n, m, k), hash_t(metro, i, j, n, m, k)};
                     fin_clauses.push_back(res);
                     res = {-hash_d(metro, i+1, j, 0, n, m, k), -hash_d(metro, i, j, 3, n, m, k), hash_t(metro, i, j, n, m, k)};         
@@ -125,7 +125,7 @@ void limit_on_turns(long long n , long long m , long long k , long long J , vect
                 }
                 if(j-1>=0){
                     // (x,y-1,R)^(x,y,U)=>T(metro,x,y) , (x,y-1,R)^(x,y,D)=>T(metro,x,y)
-                    vector<long long> res;
+                    vector<int> res;
                     res = {-hash_d(metro, i, j-1, 1, n, m, k), -hash_d(metro, i, j, 0, n, m, k), hash_t(metro, i, j, n, m, k)};
                     fin_clauses.push_back(res);
                     res = {-hash_d(metro, i, j-1, 1, n, m, k), -hash_d(metro, i, j, 2, n, m, k), hash_t(metro, i, j, n, m, k)};
@@ -133,7 +133,7 @@ void limit_on_turns(long long n , long long m , long long k , long long J , vect
                 }
                 if(j+1<m){
                     // (x,y+1,L)^(x,y,U)=>T(metro,x,y) , (x,y+1,L)^(x,y,D)=>T(metro,x,y)        
-                    vector<long long> res;
+                    vector<int> res;
                     res = {-hash_d(metro, i, j+1, 3, n, m, k), -hash_d(metro, i, j, 0, n, m, k), hash_t(metro, i, j, n, m, k)};
                     fin_clauses.push_back(res);
                     res = {-hash_d(metro, i, j+1, 3, n, m, k), -hash_d(metro, i, j, 2, n, m, k), hash_t(metro, i, j, n, m, k)};
@@ -145,13 +145,13 @@ void limit_on_turns(long long n , long long m , long long k , long long J , vect
         encode(variables, fin_clauses, n, m, k, J, seed+metro);
     }
 }
-void at_most_one_p(long long n, long long m, long long k, vector<vector<long long>>& fin_clauses) {
+void at_most_one_p(int n, int m, int k, vector<vector<int>>& fin_clauses) {
     for(int i=0; i<n; i++){
         for(int j=0; j<m; j++){
-            vector<long long> variables;
+            vector<int> variables;
             for(int metro=0; metro<k; metro++){
                 for(int metro2 = metro+1; metro2<k; metro2++){
-                    vector<long long> res;
+                    vector<int> res;
                     res = {-hash_p(metro, i, j, n, m, k), -hash_p(metro2, i, j, n, m, k)};
                     fin_clauses.push_back(res);
                 }
@@ -159,7 +159,7 @@ void at_most_one_p(long long n, long long m, long long k, vector<vector<long lon
         }
     }
 }
-void path_consistency(long long n, long long m, long long k, vector<vector<long long>>& fin_clauses) {
+void path_consistency(int n, int m, int k, vector<vector<int>>& fin_clauses) {
     for(int metro=0; metro<k; metro++){
         for(int i=0; i<n; i++){
             for(int j=0; j<m; j++){
@@ -167,7 +167,7 @@ void path_consistency(long long n, long long m, long long k, vector<vector<long 
                     fin_clauses.push_back({-hash_d(metro, i, j, 1, n, m, k)});
                 }
                 else{
-                    vector<long long> res;
+                    vector<int> res;
                     res.push_back(-hash_d(metro, i, j, 1, n, m, k));
                     res.push_back(hash_p(metro, i, j+1, n, m, k));
                     fin_clauses.push_back(res); 
@@ -180,7 +180,7 @@ void path_consistency(long long n, long long m, long long k, vector<vector<long 
                     fin_clauses.push_back({-hash_d(metro, i, j, 3, n, m, k)});
                 }
                 else{
-                    vector<long long> res;
+                    vector<int> res;
                     res.push_back(-hash_d(metro, i, j, 3, n, m, k));
                     res.push_back(hash_p(metro, i, j-1, n, m, k));
                     fin_clauses.push_back(res); 
@@ -193,7 +193,7 @@ void path_consistency(long long n, long long m, long long k, vector<vector<long 
                     fin_clauses.push_back({-hash_d(metro, i, j, 2, n, m, k)});
                 }
                 else{
-                    vector<long long> res;      
+                    vector<int> res;      
                     res.push_back(-hash_d(metro, i, j, 2, n, m, k));
                     res.push_back(hash_p(metro, i+1, j, n, m, k));
                     fin_clauses.push_back(res); 
@@ -205,7 +205,7 @@ void path_consistency(long long n, long long m, long long k, vector<vector<long 
                     fin_clauses.push_back({-hash_d(metro, i, j, 0, n, m, k)});
                 }
                 else{
-                    vector<long long> res;      
+                    vector<int> res;      
                     res.push_back(-hash_d(metro, i, j, 0, n, m, k));
                     res.push_back(hash_p(metro, i-1, j, n, m, k));
                     fin_clauses.push_back(res); 
@@ -217,32 +217,32 @@ void path_consistency(long long n, long long m, long long k, vector<vector<long 
         }
     }
 }
-void continuity(long long n , long long m , long long t , vector<vector<long long>>& clauses , set<pair<int,int>>& end_points) {
+void continuity(int n , int m , int t , vector<vector<int>>& clauses , set<pair<int,int>>& end_points) {
     for(int i = 1 ; i <= n; i++) {
         for(int j = 1; j <= m; j++) {
             if(end_points.find({i,j}) != end_points.end()) continue;
             for(int k = 0; k < t; k++) {
                 for(int l = 5 ; l <= 8; l++) {
                     if(i - 1 >= 1 && l == 5) {
-                        vector<long long>res;
+                        vector<int>res;
                         res.push_back(-1LL*((i*m + j - 1)*(8LL*t) + 8*k + l));
                         res.push_back(( ( (i-1)*m + j - 1)*(8LL*t) + 8*k + 3));
                         clauses.push_back(res);
                     }
                     if(j + 1 <= m && l == 6) {
-                        vector<long long>res;
+                        vector<int>res;
                         res.push_back(-1LL*((i*m + j - 1)*(8LL*t) + 8*k + l));
                         res.push_back(((i*m + (j+1) - 1)*(8LL*t) + 8*k + 4));
                         clauses.push_back(res);
                     }
                     if(i + 1 <= n && l == 7) {
-                        vector<long long>res;
+                        vector<int>res;
                         res.push_back(-1LL*((i*m + j - 1)*(8LL*t) + 8*k + l));
                         res.push_back((((i+1)*m + j - 1)*(8LL*t) + 8*k + 1));
                         clauses.push_back(res);
                     }
                     if(j - 1 >= 1 && l == 8) {
-                        vector<long long>res;
+                        vector<int>res;
                         res.push_back(-1LL*((i*m + j - 1)*(8LL*t) + 8*k + l));
                         res.push_back(((i*m + (j - 1) - 1)*(8LL*t) + 8*k + 2));
                         clauses.push_back(res);
@@ -252,7 +252,7 @@ void continuity(long long n , long long m , long long t , vector<vector<long lon
         }
     }
 }
-void create_inputfile_for_sat(vector<vector<long long>>& clauses, string filename) {
+void create_inputfile_for_sat(vector<vector<int>>& clauses, string filename) {
     ofstream myfile;
     myfile.open (filename);
     myfile << "p cnf " << 1000000000 << " " << clauses.size() << "\n";
